#107: use /packages route; package comparison is related to popularity so putting it together makes sense
